requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
filelock==3.9.0
Flask==2.2.2
ftfy==6.1.1
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
